Split replicator
