+++ src/lib.rs
@@ -44,7 +44,7 @@ pub mod fleece_mutable;
 pub mod index;
 pub mod logging;
 pub mod query;
-pub mod replicator;
+pub mod replication;
 pub mod slice;
 
 mod c_api;
@@ -67,7 +67,7 @@ pub use error::*;
 pub use fleece::*;
 pub use fleece_mutable::*;
 pub use query::*;
-pub use replicator::*;
+pub use replication::*;
 
 //////// TOP-LEVEL TYPES:
