+++ tests/document_tests.rs
@@ -2,6 +2,7 @@ extern crate core;
 extern crate couchbase_lite;
 
 use self::couchbase_lite::*;
+use self::couchbase_lite::callbacks::*;
 use std::time::Duration;
 
 pub mod utils;
