+++ tests/replicator_tests.rs
@@ -18,6 +18,8 @@
 extern crate couchbase_lite;
 
 use self::couchbase_lite::*;
+use self::couchbase_lite::callbacks::*;
+use self::couchbase_lite::configuration::*;
 use encryptable::Encryptable;
 use std::{time::Duration, thread};
